requirements.txt
discord.py
twitchAPI
python-dotenv
